More complete overflow fix
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/6088


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@23827 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/smpdtfmt.cpp
@@ -686,7 +686,7 @@ SimpleDateFormat::parseGMT(const UnicodeString &text, ParsePosition &pos) const 
                 if (pos.getErrorIndex() == -1 && pos.getIndex() > start) {
                     parsed.getArray(parsedCount);
                     if (parsedCount == 1 && parsed[0].getType() == Formattable::kDate) {
-                        return (int32_t)(-1 * parsed[0].getDate());
+                        return (int32_t)(-1 * (int64_t)parsed[0].getDate());
                     }
                 }
 
@@ -699,7 +699,7 @@ SimpleDateFormat::parseGMT(const UnicodeString &text, ParsePosition &pos) const 
                 if (pos.getErrorIndex() == -1 && pos.getIndex() > start) {
                     parsed.getArray(parsedCount);
                     if (parsedCount == 1 && parsed[0].getType() == Formattable::kDate) {
-                        return (int32_t)parsed[0].getDate();
+                        return (int32_t)((int64_t)parsed[0].getDate());
                     }
                 }
 
@@ -725,7 +725,7 @@ SimpleDateFormat::parseGMT(const UnicodeString &text, ParsePosition &pos) const 
                 if (pos.getErrorIndex() == -1 && pos.getIndex() > start) {
                     parsed.getArray(parsedCount);
                     if (parsedCount == 1 && parsed[0].getType() == Formattable::kDate) {
-                        return (int32_t)parsed[0].getDate();
+                        return (int32_t)((int64_t)parsed[0].getDate());
                     }
                 }
